added progress bar while passing course & some code style improvements
