+++ app/assets/stylesheets/application.scss
@@ -23,6 +23,7 @@
 .navbar-brand{
   margin-left: 10%;
 }
+
 .fa.fa-check-circle.fa-2x{
   margin-right: 10px;
   margin-top: 2px;
@@ -34,19 +35,23 @@
   margin-top: 2px;
   margin-right: 10px;
 }
+
 .fa.fa-pencil.fa-2x{
   color: blue;
   margin-top: 2px;
   margin-right: 10px;
 }
+
 .fa.fa-trash.fa-2x{
   color: red;
   margin-top: 2px;
   margin-right: 10px;
 }
+
 .title.is-2.center{
   text-align: center;
 }
+
 .tag.is-info.is-medium.is-custom{
   margin-right: 20px;
 }
@@ -58,9 +63,11 @@
 #checkbox-from{
   display: none;
 }
+
 #q-radio-from{
   display: none;
 }
+
 #answer{
   margin-left: 0px;
 }
@@ -75,6 +82,7 @@
   width: 128px;
   margin: 0 auto;
 }
+
 .edit_relationship{
   width: 128px;
   margin: 0 auto;
@@ -83,13 +91,16 @@
 #rate-bnt{
   display: none;
 }
+
 .ff{
   width: 240px;
   margin-left: 72px;
 }
+
 #answers_{
   width: 163px;
 }
+
 .button.is-small.is-primary.quest{
   margin-left: 105px; 
   margin-top: 5px; 
@@ -103,7 +114,26 @@
   margin-left: 103px; 
   margin-top: 5px; 
 }
+
 .t-quest{
   width: 307px;
   margin-left: 70px;
+}
+
+.ui-progressbar {
+  height: 1em;
+  text-align: left;
+  overflow: hidden;
+  width: 80%;
+  margin: 0 auto;
+  background: #e9ecef;
+  border-radius: 290486px;
+}
+
+.ui-progressbar .ui-progressbar-value {
+  margin: -1px;
+  height: 100%;
+  background-image: linear-gradient(45deg,rgba(255,255,255,.15) 25%,transparent 25%,transparent 50%,rgba(255,255,255,.15) 50%,rgba(255,255,255,.15) 75%,transparent 75%,transparent);
+  background-size: 1rem 1rem;
+  background-color: #ffc107;
 }
