+++ app/helpers/user_course_helper.rb
@@ -1,0 +1,6 @@
+module UserCourseHelper
+  
+ def check_on_completed(current_user)
+    current_user.user_courses.pluck(:course_id).include?(@course.id)
+  end
+end
