+++ app/views/pages/show.html.erb
@@ -1,3 +1,4 @@
+<div id="progressbar"></div>
 <div class="columns is-centered">
   <div class="column is-2">
     <h1 class="title is-2"><%= @page.title %></h1>
@@ -16,3 +17,11 @@
     <%= link_to 'Back', course_path(@course),{ class: "button  is-small is-black"}  %>
   </div>
 </div>
+
+<script type="text/javascript">
+  $(document).ready(function(){
+    $('#progressbar').progressbar({
+      value: <%= percent_of_progess(@course, current_user) %>
+    });
+  });
+</script>
