+++ app/views/user_courses/_complete_course_form.html.erb
@@ -1,5 +1,5 @@
 <% if all_user_pages_completed? && !check_on_completed(current_user) %>
-  <%= form_tag complete_user_course_path(@course), method: :patch, remote: true,  class: 'render-form' do %>
+  <%= form_tag complete_user_course_path(@course), method: :patch, class: 'render-form' do %>
     <%= hidden_field_tag(:user_id, current_user.id) %>
     <%= hidden_field_tag(:course_id, @course.id) %>
     <%= hidden_field_tag(:completed, true)%>
